requirements.txt
certifi==2018.8.24
chardet==3.0.4
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.3
decorator==4.3.0
dill==0.2.8.2
EMD-signal==0.2.5
future==0.16.0
idna==2.7
kiwisolver==1.0.1
line-bot-sdk==1.8.0
matplotlib==3.0.0
multiprocess==0.70.6.1
networkx==2.2
numpy==1.15.2
pandas==0.23.4
pathos==0.2.2.1
Pillow==5.3.0
pkg-resources==0.0.0
pox==0.2.4
ppft==1.6.4.8
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
requests==2.19.1
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
toolz==0.9.0
torch==0.4.1
tqdm==4.27.0
urllib3==1.23
